requirements.txt
anthropic[bedrock]>=0.40.0
PyGithub>=2.1.0
